Backend/requirements.txt
Flask==2.3.3
requests==2.32.0
python-dotenv
pylint